day -3 creation and deletion done
creation and deletion of courses has been done and associations have been given to the tables . we have to serialize user and try to add creator id based on the userId. in progess . however the design seems to be boring and not engaging . should research about it and should make it attractive enough. should add authentication and display of username so that the profile page is completed .. this is later work .. the focus is now on the creator side of the personas:creators..